+++ libytk/ytk.h
@@ -37,15 +37,15 @@ typedef struct __ytk_thing {
 	ylong	top, left;
 	char	*title;
 	void	(*escape) (struct __ytk_thing *);
-	int		colors;
-	int		attr;
+	long unsigned int colors;
+	long unsigned int attr;
 } ytk_thing;
 
 ytk_thing *ytk_new_thing(void);
 void ytk_delete_thing(ytk_thing *);
 void ytk_set_escape(ytk_thing *, void (*) (ytk_thing *));
-void ytk_set_colors(ytk_thing *, int);
-void ytk_set_attr(ytk_thing *, int);
+void ytk_set_colors(ytk_thing *, long unsigned int);
+void ytk_set_attr(ytk_thing *, long unsigned int);
 
 #include "stack.h"
 #include "menu.h"
